Refactor orikaconfiguration to domain module. Move mapper tests back.
