+++ oppijanumerorekisteri-domain/src/test/java/fi/vm/sade/oppijanumerorekisteri/mappers/HenkiloMapperTest.java
@@ -9,7 +9,6 @@ import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
-import fi.vm.sade.oppijanumerorekisteri.configurations.OrikaConfiguration;
 
 import java.util.Date;
