+++ oppijanumerorekisteri-service/src/main/java/fi/vm/sade/oppijanumerorekisteri/services/impl/HenkiloServiceImpl.java
@@ -1,7 +1,7 @@
 package fi.vm.sade.oppijanumerorekisteri.services.impl;
 
 import com.querydsl.core.types.Predicate;
-import fi.vm.sade.oppijanumerorekisteri.configurations.OrikaConfiguration;
+import fi.vm.sade.oppijanumerorekisteri.mappers.OrikaConfiguration;
 import fi.vm.sade.oppijanumerorekisteri.dto.*;
 import fi.vm.sade.oppijanumerorekisteri.exceptions.NotFoundException;
 import fi.vm.sade.oppijanumerorekisteri.exceptions.UserHasNoOidException;
